filter details

+++ src/orders/orders.controller.ts
@@ -7,6 +7,11 @@ import {
   Param,
   Delete,
   UseGuards,
+  ForbiddenException,
+  ParseBoolPipe,
+  Query,
+  DefaultValuePipe,
+  ParseIntPipe,
 } from '@nestjs/common';
 import { OrdersService } from './orders.service';
 import { CreateOrderDto } from './dto/create-order.dto';
@@ -14,7 +19,8 @@ import { UpdateOrderDto } from './dto/update-order.dto';
 import { AtGuard, RolesGuard } from 'src/auth/guards';
 import { Role } from 'src/users/enums/user-role.enum';
 import { Roles } from 'src/auth/decorators';
-import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiQuery, ApiTags } from '@nestjs/swagger';
+import { UserD } from 'src/auth/decorators/users.decorator';
 
 @ApiTags('orders')
 @ApiBearerAuth()
@@ -35,10 +41,31 @@ export class OrdersController {
     return this.ordersService.findAll();
   }
 
-  @Roles(Role.ADMIN)
+  @Roles(Role.ADMIN, Role.USER)
+  @ApiQuery({
+    name: 'details',
+    required: false,
+    type: 'boolean',
+    default: false,
+    description: 'Get user details with more info ',
+  })
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.ordersService.findOne(+id);
+  async findOne(
+    @Param('id', ParseIntPipe) id: number,
+    @UserD('sub') token_id: number,
+    @UserD('role') token_role: Role,
+    @Query('details', new DefaultValuePipe(false), ParseBoolPipe)
+    details?: boolean,
+  ) {
+    if (token_id !== id && token_role !== Role.ADMIN) {
+      throw new ForbiddenException('You are not authorized to get this order');
+    }
+    try {
+      return await this.ordersService.findOne(id, details);
+    } catch (error) {
+      const err = error as Error;
+      throw new Error(err.message);
+    }
   }
 
   @Roles(Role.ADMIN, Role.USER)
@@ -47,9 +74,16 @@ export class OrdersController {
     return this.ordersService.update(+id, updateOrderDto);
   }
 
-  @Roles(Role.ADMIN)
+  @Roles(Role.ADMIN, Role.USER)
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(
+    @Param('id', ParseBoolPipe) id: number,
+    @UserD('sub') token_id: number,
+    @UserD('role') token_role: Role,
+  ) {
+    if (token_id !== id && token_role !== Role.USER) {
+      throw new ForbiddenException('You are not authorized to delete orders');
+    }
     return this.ordersService.remove(+id);
   }
 }
